ci: run the git-daemon integration tests on Windows too (Phase 2b)
Closes the last Windows gap. The job now runs every fastskill-core test --
657, up from 637 -- with no exclusion filter at all.

The three git-daemon files were excluded because their fixture killed the
daemon by walking /proc/<pid>/task/<pid>/children, which does not exist on
Windows. That fixture was also triplicated across the three files, so it is
now extracted to tests/common/mod.rs and ported once: Linux keeps the
/proc descendant walk, Windows uses `taskkill /F /T /PID` to kill the tree,
and anything else falls back to a plain kill.

Reaping is now bounded rather than a blocking wait(). If a kill somehow
fails to land, the old code would hang the test -- and so the whole suite --
instead of failing it. Worst case is now ~10s and a clear panic.

Nothing is #[cfg(unix)]-gated: every platform gets a working kill path and
no test is compiled out.

What this finally covers on Windows: the cache's integration behaviour, and
the #[cfg(windows)] rename-retry branch in core::cache, which has compiled
since Phase 1 but until now had nothing exercising it.
