+++ .github/workflows/test.yml
@@ -46,7 +46,7 @@ jobs:
       - run: cargo install cargo-nextest --locked
       - run: cargo nextest run --retries 3 --fail-fast -E 'not test(install_e2e_tests)'
 
-  # Windows unit + integration test coverage (Phase 2a).
+  # Windows unit + integration test coverage (Phase 2a + 2b).
   #
   # release.yml ships an x86_64-pc-windows-msvc binary, but every other job
   # here is ubuntu-only, so Windows-specific code has never executed anywhere.
@@ -55,18 +55,15 @@ jobs:
   # open) has never even been compiled -- it cannot be built from Linux
   # because `ring` and `onig_sys` need MSVC tooling to cross-compile.
   #
-  # Scope is the fastskill-core library unit tests (Phase 1) plus the
-  # integration tests under crates/fastskill-core/tests/, with default
-  # features. Excluded are the three integration-test binaries that spin up
-  # a real `git daemon`. Their fixture is unported rather than known-broken:
-  # it kills the daemon by finding its child process through
-  # /proc/<pid>/task/<pid>/children, which does not exist on Windows. (Git
-  # for Windows does ship git-daemon, so these may well work once the
-  # fixture uses a portable process-tree kill.) The three files are:
-  #   - git_content_cache_test
-  #   - offline_verification_sweep_test
-  #   - repo_index_refresh_git_test
-  # Porting those is Phase 2b.
+  # Scope is the fastskill-core library unit tests plus every integration
+  # test under crates/fastskill-core/tests/, with default features -- the
+  # same real `git daemon` fixture (tests/common/mod.rs) that ubuntu uses,
+  # including the three binaries that spin one up
+  # (git_content_cache_test, offline_verification_sweep_test,
+  # repo_index_refresh_git_test). Their only prior Windows blocker was
+  # killing the daemon's process tree via /proc/<pid>/task/<pid>/children,
+  # which does not exist on Windows; the fixture now uses `taskkill /F /T`
+  # there instead (Phase 2b).
   test-windows:
     name: Test (Windows, core unit + integration tests)
     runs-on: windows-latest
@@ -76,7 +73,7 @@ jobs:
       - run: cargo install cargo-nextest --locked
       # --no-fail-fast: this job is still being brought up, and stopping at the
       # first failure would surface only one platform issue per CI round-trip.
-      - run: cargo nextest run -p fastskill-core --lib --test '*' --retries 3 --no-fail-fast -E 'not binary(git_content_cache_test) and not binary(offline_verification_sweep_test) and not binary(repo_index_refresh_git_test)'
+      - run: cargo nextest run -p fastskill-core --lib --test '*' --retries 3 --no-fail-fast
 
   test-all-features:
     name: Test (All Features)
